This commit was manufactured by cvs2svn to create tag 'Ext-5-1-1-pre2'.
git-svn-id: file:///home/hardaker/lib/sf-bkups/net-snmp-convert-svnrepo/tags/Ext-5-1-1-pre2@9708 06827809-a52a-0410-b366-d66718629ded

+++ net-snmp/agent/mibgroup/utilities/execute.c
@@ -71,11 +71,11 @@ run_shell_command( char *command, char *input,
                     unlink(ifname);
                 return -1;
             }
-            snprintf( shellline, sizeof(shellline), "(%s) < %s > %s",
+            snprintf( shellline, sizeof(shellline), "(%s) < \"%s\" > \"%s\"",
                       command, ifname, ofname );
         } else {
             ofname = NULL;   /* Just to shut the compiler up! */
-            snprintf( shellline, sizeof(shellline), "(%s) < %s",
+            snprintf( shellline, sizeof(shellline), "(%s) < \"%s\"",
                       command, ifname );
         }
     } else {
@@ -84,7 +84,7 @@ run_shell_command( char *command, char *input,
             ofname = make_tempfile();
             if(NULL == ofname)
                 return -1;
-            snprintf( shellline, sizeof(shellline), "(%s) > %s",
+            snprintf( shellline, sizeof(shellline), "(%s) > \"%s\"",
                       command, ofname );
         } else {
             ofname = NULL;   /* Just to shut the compiler up! */
@@ -107,9 +107,11 @@ run_shell_command( char *command, char *input,
         int         len = 0;
         fd = open(ofname, O_RDONLY);
         if(fd >= 0)
-            len  = read( fd, output, *out_len );
+            len  = read( fd, output, *out_len-1 );
 	*out_len = len;
-	close(fd);
+	if (len >= 0) output[len] = 0;
+	else output[0] = 0;
+	if (fd >= 0) close(fd);
         unlink(ofname);
     }
     if ( input ) {
@@ -187,6 +189,7 @@ int
 run_exec_command( char *command, char *input,
                   char *output,  int  *out_len)	/* Or realloc style ? */
 {
+#if HAVE_EXECV
     int ipipe[2];
     int opipe[2];
     int i, len;
@@ -195,7 +198,6 @@ run_exec_command( char *command, char *input,
     char **argv;
     int argc;
 
-#if HAVE_EXECV
     pipe(ipipe);
     pipe(opipe);
     if ((pid = fork()) == 0) {
@@ -245,17 +247,19 @@ run_exec_command( char *command, char *input,
 	   write(ipipe[1], input, strlen(input));
 	   close(ipipe[1]);	/* or flush? */
         }
+	else close(ipipe[1]);
         if (waitpid(pid, &result, 0) < 0 ) {
             snmp_log_perror("waitpid");
             return -1;
         }
         if (output) {
-            len = read( opipe[0], output, *out_len );
+            len = read( opipe[0], output, *out_len-1 );
+	    if (len >= 0) output[len] = 0;
+	    else output[0] = 0;
 	    *out_len = len;
         }
-	close(ipipe[1]);
 	close(opipe[0]);
-	return 0;
+	return WEXITSTATUS(result);
 
     } else {
         /*
